The Great <Div>ide
Now that I’m not using JavaScript to hand the switch between GR/EN but
am linking two separate documents, everything gets a lot cleaner.

I got rid of unnecessary <div>’s for the “oververses” and introduced
the classes “gr-underverse, eng-underverse” which are toggle-able.

The style sheet is still a mile long, but this week is all about
consolidation.

Priorities:

1. Use PHP to clean up markdown.
2. Eliminate redundancies in the style.
3. Add semi-trans verse numbers.
4. Reintroduce JS/JQ elements on page transitions.
5. Finish breaking up and dropping in the verses for Book I.