Ordered Listing
With a swamp of thing going on recently, I made a very obvious mistake.
I didn’t use an ordered list for my verse number.

Sometimes the Greek and English don’t match up exactly, so I wanted the
freedom to mark them as need (hence my <span .num> solution>.

But it occurred to me anytime I need to make an exception it’d be easy
to tack on an empty <li></li> to keep things even.

Now the verse numbers are looking schnazzy.